+++ Source/Shields/IShield.cs
@@ -9,6 +9,6 @@ namespace FrontierDevelopments.Shields
         bool Collision(Vector3 point);
         Vector3? Collision(Ray ray, float limit);
         Vector3? Collision(Vector3 start, Vector3 end);
-        bool Block(int damage, Vector3 position);
+        bool Block(long damage, Vector3 position);
     }
 }
